+++ CHANGELOG.md
@@ -14,6 +14,9 @@
   - Added rails-controller-testing
 
 ### Refactoring
+- Rank the Research functionality has been removed
+  - Accepted research articles are available on the blog, and suggested research
+    topics can now be discussed on the forum
 - Cleaned up existing migration files
 - Simplified configuration options
 - All emails are now sent in a background process